requirements.txt
neo4j>=5.28.0
python-dotenv>=1.0.1
openai>=1.60.0
pydantic>=2.10.4
rapidfuzz>=3.11.0
tqdm>=4.67.1
google-auth>=2.0.0
requests>=2.0.0
pypdf>=5.0.0

